refactor: 改进代码质量和修复多处问题
修复内容:
- 修复 MiniMindConfig 中 rope_theta 类型注解错误 (int -> float)
- 修复 trainer_utils.py 中运算符空格问题 (from_weight!= -> from_weight !=)
- 移除 model_lora.py 中未使用的 optim 导入
- 修复多个训练脚本中 max_seq_len 参数问题，统一使用 max_position_embeddings
- 更新弃用的 torch.cuda.amp API 为新的 torch.amp API (autocast 和 GradScaler)
- 为 trainer_utils.py 和 model_lora.py 添加类型提示和文档字符串
- 改进代码一致性和可读性

+++ model/model_minimind.py
@@ -22,7 +22,7 @@ class MiniMindConfig(PretrainedConfig):
             num_key_value_heads: int = 2,
             vocab_size: int = 6400,
             rms_norm_eps: float = 1e-05,
-            rope_theta: int = 1000000.0,
+            rope_theta: float = 1000000.0,
             inference_rope_scaling: bool = False,
             flash_attn: bool = True,
             ####################################################
